Use special rule for BlockedByKing
Simplification: remove BlockedByKing from storm array and use a special
rule.

The BlockedByKing section in the storm array is substantially similar to
the
Unopposed section except for two extreme values V(-290), V(-274). Turns
out
removing BlockedByKing and using a special rule for these two values
shows
no Elo loss. All the other values in the BlockedByKing section are
apparently
irrelevant. BlockedByKing now falls under unopposed which (to me) is a
bit
more logical since there is no defending pawn on this file. Also,
retuning
the Unopposed section may be another improvement.

GOOD) This is a simplification because the entire BlockedByKing section
of
the storm array goes away reducing a few lines of code (and less values
to
tune). This also brings clarity because the special rule is self
documenting.

BAD) It takes execution time to apply the special rule. This should be
negli-
gible because it is based on a template parameter and is boiled down to
two
bitwise AND's.

STC:
LLR: 2.96 (-2.94,2.94) [-3.00,1.00]
Total: 33470 W: 6820 L: 6721 D: 19929
http://tests.stockfishchess.org/tests/view/5ae7b6e60ebc5926dba90e13

LTC:
LLR: 2.96 (-2.94,2.94) [-3.00,1.00]
Total: 47627 W: 7045 L: 6963 D: 33619
http://tests.stockfishchess.org/tests/view/5ae859ff0ebc5926dba90e85

Closes #1574

Bench: 5512000

-----------

How to continue after this patch?

This patch may open the possibility to move the special rule to
evaluate.cpp
in the evaluate::king() function, where we could refine the rule using
king
danger information. For instance, with a king in H2 blocking an opponent
pawn
in H3, it may be critical to know that the opponent has no safe check in
G2
before giving the bonus :-)
master (#11)
@protonspring
@snicolet
protonspring authored