Tweak kingAttackersCount and KingAttackWeights
Use the whole kingRing for pawn attackers instead of only the squares
directly
around the king. This tends to give quite a lot more kingAttackersCount,
so to
compensate and to avoid raising the  king danger too fast we lower the
values
in the KingAttackWeights array a little bit.

STC:
LLR: 2.95 (-2.94,2.94) [0.00,4.00]
Total: 51892 W: 10723 L: 10369 D: 30800
http://tests.stockfishchess.org/tests/view/5af6d4dd0ebc5968e652428e

LTC:
LLR: 2.96 (-2.94,2.94) [0.00,4.00]
Total: 24536 W: 3737 L: 3515 D: 17284
http://tests.stockfishchess.org/tests/view/5af709890ebc5968e65242ac

Credits to user @xoroshiro for the idea of using the kingRing for pawn
attackers.

How to continue? It seems that the KingAttackWeights[] array stores
values
which are quite Elo-sensitive, yet they have not been tuned with SPSA
recently.
There might be easy Elo points to get there.

Closes #1597

Bench: 5282815
master
@snicolet
snicolet committed