Simplify BlockedByPawn to one dimension
I was able to get this to pass which reduces BlockedByPawn to one
dimension
with NO distance from edge offset.

GOOD)  It's more simple and may provide additional clarity for further
simplifications.  Facilitates migrating unblocked to one dimension as
well.

BAD) If there is indeed a distance component to BlockedStorm (may or may
not be the case), this obfuscates this component into ShelterStrength
and
UnblockedStorm. This may be more convoluted. Also, it may be more
convenient
to have each of the three arrays (ShelterStrength, BlockedStorm, and
UnBlocked)
be the same size.

STC:
LLR: 2.96 (-2.94,2.94) [-3.00,1.00]
Total: 96173 W: 19326 L: 19343 D: 57504
http://tests.stockfishchess.org/tests/view/5b04544d0ebc5914abc12965

LTC:
LLR: 2.95 (-2.94,2.94) [-3.00,1.00]
Total: 49818 W: 7441 L: 7363 D: 35014
http://tests.stockfishchess.org/tests/view/5b0487d50ebc5914abc12990

Closes #1611

Bench: 5133208
master (#2)
@protonspring
@snicolet
protonspring authored