Use threads to clear the TT faster.
Stockfish currently takes a while to clear the TT when using larger hash
sizes.

On one machine with 128 GB hash it takes about 50 seconds with a single
thread,
allowing it to use all allocated cores brought that time down to 4
seconds on
some Linux systems. The patch was further tested on Windows and refined
with
NUMA binding of the hash initializing threads (we refer to pull request
#1601
for the complete discussion and the speed measurements).

Closes #1601

No functional change
master
@RodrigoTjader
@snicolet
RodrigoTjader authored